Refactor : Cam 컴포넌트 일부 이름 변경
- 다른 page 들과의 일관성을 위해 Cam을 CamMain으로 변경
- sharedScreen class는 component가 아니므로 /utils로 이동
