+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { RecoilRoot } from 'recoil';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 import LoginMain from './components/LoginPage/LoginMain';
-import Cam from './components/Cam/Cam';
+import CamMain from './components/Cam/CamMain';
 import BoostCamMain from './components/Main/BoostCamMain';
 import LoginCallback from './components/LoginPage/LoginCallback';
 
@@ -15,7 +15,7 @@ function App(): JSX.Element {
           <Route path="/" element={<LoginMain />} />
           <Route path="/main" element={<BoostCamMain />} />
           <Route path="/login/github" element={<LoginCallback service="github" />} />
-          <Route path="/cam" element={<Cam />} />
+          <Route path="/cam" element={<CamMain />} />
         </Routes>
       </RecoilRoot>
     </Router>
